% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_4j.R
\name{download_4j}
\alias{download_4j}
\title{Download Neo4J community}
\usage{
download_4j(path = "internal", flavor = c("unix", "windows"))
}
\arguments{
\item{path}{where to download Neo4J. Default is in the package folder.}

\item{flavor}{wether to download for unix or windows.}
}
\value{
an installed Neo4J
}
\description{
Download Neo4J community
}
