% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{neo4j_api}
\alias{neo4j_api}
\title{Neo4J API}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
neo4j_api
}
\value{
An R6 object that can call the API
}
\description{
Neo4J API
}
\keyword{datasets}
