% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{neo4j_shell}
\alias{neo4j_shell}
\title{Run Neo4J}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
neo4j_shell
}
\value{
An R6 object that can manage Neo4J
}
\description{
Run Neo4J
}
\examples{
\dontrun{
db <- neo4j_shell$new(neo4jhome = "inst")
db$start()
db$see_logs(size = 10)
db$version()
db$restart()
db$cypher(cypher = "MATCH (n) RETURN COUNT(*);",
          username = "neo4j", password = "pouetpouet", format = "verbose")
}
}
\keyword{datasets}
